Add files via upload
This is my very first project built entirely with HTML and CSS. I'm just getting started in Front-End development and I'm open to tips and constructive feedback. I followed a course where we recreated the Starbucks website using only HTML and CSS, and this project is the result. I'm now moving on to JavaScript and planning to develop more personal projects with it.